+++ nacc_attribute_deriver/attributes/nacc/modules/__init__.py
@@ -1,5 +1,6 @@
 # ruff: noqa
 from .cls import *
+from .csf import *
 from .ftld import *
 from .lbd import *
 from .cross_module import *