+++ nacc_attribute_deriver/utils/scope.py
@@ -11,6 +11,7 @@ class Scope(str, Enum):
 class FormScope(Scope):
     # forms
     CLS = "cls"
+    CSF = "csf"
     MDS = "mds"
     MILESTONE = "milestone"
     NP = "np"
@@ -45,11 +46,12 @@ class SCANMRIScope(Scope):
 
 class ImagingScope(Scope):
     # MP scope
-    MIXED_PROTOCOL = "mp"
+    MRI_SUMMARY = "mri_summary"
 
 
 ScopeLiterals = Literal[
     FormScope.CLS,
+    FormScope.CSF,
     FormScope.MDS,
     FormScope.MILESTONE,
     FormScope.NP,
@@ -68,5 +70,5 @@ ScopeLiterals = Literal[
     SCANPETScope.TAU_PET_NPDKA,
     SCANMRIScope.MRI_QC,
     SCANMRIScope.MRI_SBM,
-    ImagingScope.MIXED_PROTOCOL,
+    ImagingScope.MRI_SUMMARY,
 ]